refactor node location state
