+++ nym-vpn-x/src-tauri/src/commands/mod.rs
@@ -5,5 +5,5 @@ pub mod daemon;
 pub mod db;
 pub mod fs;
 pub mod log;
-pub mod node_location;
+pub mod country;
 pub mod window;
