+++ nym-vpn-x/src-tauri/src/main.rs
@@ -21,6 +21,7 @@ use commands::db as cmd_db;
 use commands::fs as cmd_fs;
 use commands::log as cmd_log;
 use commands::window as cmd_window;
+use commands::country as cmd_country;
 use commands::*;
 use nym_config::defaults;
 use states::app::AppState;
@@ -203,9 +204,7 @@ async fn main() -> Result<()> {
             cmd_db::db_set,
             cmd_db::db_get,
             cmd_db::db_flush,
-            node_location::get_node_location,
-            node_location::set_node_location,
-            node_location::get_countries,
+            cmd_country::get_countries,
             cmd_window::show_main_window,
             commands::cli::cli_args,
             cmd_log::log_js,
